Componentize the 'tablature grill' -- the text area with | values that sit infront of the TablatureBlocl
